requirements.txt
aniso8601==1.2.1
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
Flask==0.11.1
Flask-Cors==3.0.2
flask-restplus==0.10.1
google-api-python-client==1.6.2
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
oauth2client==4.1.2
plaid-python==2.0.1
py==1.4.34
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.18
PyJWT==1.5.2
pymongo==3.4.0
pyOpenSSL==17.1.0
PySocks==1.6.7
pytest==3.1.3
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.1
requests-mock==1.3.0
rsa==3.4.2
six==1.10.0
twilio==6.0.0
uritemplate==3.0.0
urllib3==1.21.1
Werkzeug==0.12.2
